requirements.txt

# Production dependencies
psycopg2-binary>=2.9.9
Pillow>=10.0.0
bleach>=6.1.0
celery>=5.3.0
channels-redis>=4.1.0
channels>=4.0.0
daphne>=4.0.0
dj-database-url==2.1.0
dj-database-url>=2.1.0
django-axes>=6.1.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
django-cors-headers>=4.3.0
django-filter>=23.0
django-ipware>=5.0.0
django-otp>=1.2.0
django-ratelimit>=4.1.0
django-redis==5.4.0
django>=4.2,<5.0
djangorestframework-simplejwt>=5.3.0
djangorestframework>=3.14.0
dnspython==2.4.2
drf-spectacular>=0.26.0
email-validator==2.1.0
eth-account>=0.10.0
gunicorn==21.2.0
gunicorn>=21.0.0
psycopg2-binary==2.9.9
psycopg2-binary>=2.9.9
pyotp>=2.9.0
python-decouple>=3.8
python-dotenv>=1.0.0
qrcode>=7.4.0
redis==5.0.1
redis>=5.0.0
web3>=6.0.0
whitenoise==6.6.0
whitenoise>=6.6.0
sentry-sdk==1.38.0
argon2-cffi==23.1.0
argon2-cffi==23.1.0
argon2-cffi==23.1.0
